refactor(banner): Address feedback
