+++ packages/react/package.json
@@ -35,7 +35,8 @@
     "type-check": "tsc"
   },
   "dependencies": {
-    "hig-vanilla": "^0.1.26"
+    "hig-vanilla": "^0.1.26",
+    "react-transition-group": "^2.3.0"
   },
   "peerDependencies": {
     "prop-types": "^15.5.10",
