+++ packages/react/src/elements/components/Banner/BannerAction/banner-action.scss
@@ -1,0 +1,42 @@
+@import "../tokens.scss";
+
+.hig__banner__action {
+  display: flex;
+  align-items: center;
+  padding: $banner-action-padding-y $banner-action-spacing-x $banner-action-padding-y 0;
+  border-right-width: 1px;
+  border-right-style: solid;
+  margin-left: $banner-action-spacing-x;
+
+  &:first-child {
+    margin-left: 0;
+  }
+
+  &:last-child {
+    .hig__banner--wrap-content & {
+      border-right: 0;
+    }
+  }
+
+  .hig__banner--urgent & {
+    border-right-color: $banner-border-color--urgent;
+  }
+
+  .hig__banner--warning & {
+    border-right-color: $banner-border-color--warning;
+  }
+
+  .hig__banner--primary & {
+    border-right-color: $banner-border-color--primary;
+  }
+
+  .hig__banner--complete & {
+    border-right-color: $banner-border-color--complete;
+  }
+
+  .hig__banner__interactions--wrap-actions & {
+    padding-right: 0;
+    border-right: 0;
+    margin-left: 0;
+  }
+}
