--- packages/react/src/elements/components/Banner/BannerAction/styles.scss
@@ -1,42 +1,0 @@
-.hig__banner__action {
-  @import "../variables.scss";
-
-  display: flex;
-  align-items: center;
-  padding: $action-padding-y $action-spacing-x $action-padding-y 0;
-  border-right-width: 1px;
-  border-right-style: solid;
-  margin-left: $action-spacing-x;
-
-  &:first-child {
-    margin-left: 0;
-  }
-
-  &:last-child {
-    .hig__banner--wrap-content & {
-      border-right: 0;
-    }
-  }
-
-  .hig__banner--error & {
-    border-right-color: $border-color--error;
-  }
-
-  .hig__banner--warning & {
-    border-right-color: $border-color--warning;
-  }
-
-  .hig__banner--primary & {
-    border-right-color: $border-color--primary;
-  }
-
-  .hig__banner--success & {
-    border-right-color: $border-color--success;
-  }
-
-  .hig__banner__interactions--wrap-actions & {
-    padding-right: 0;
-    border-right: 0;
-    margin-left: 0;
-  }
-}
