+++ packages/react/src/elements/components/Banner/BannerAnimator.js
@@ -1,7 +1,5 @@
 import PropTypes from "prop-types";
 
-import "./styles.scss";
-
 /**
  * @typedef {Object} BannerAnimatorProps
  * @property {boolean} isVisible
