--- packages/react/src/elements/components/Banner/BannerAnimator/styles.scss
@@ -1,19 +1,0 @@
-.hig__banner {
-  &--appear {
-  }
-
-  &--active-appear {
-  }
-
-  &--enter {
-  }
-
-  &--active-enter {
-  }
-
-  &--exit {
-  }
-
-  &--active-exit {
-  }
-}
