+++ packages/react/src/elements/components/Banner/BannerPresenter/banner-presenter.scss
@@ -1,0 +1,122 @@
+@import "../tokens.scss";
+
+.hig__banner {
+  display: flex;
+  box-sizing: border-box;
+  min-height: $banner-wrapper-min-height;
+  border-width: 1px 0;
+  border-style: solid;
+
+  &.hig__banner--bottom {
+    border-bottom-width: 0;
+  }
+
+  &.hig__banner--top {
+    border-top-width: 0;
+  }
+
+  &.hig__banner--urgent {
+    border-color: $banner-border-color--urgent;
+    background-color: $banner-background-color--urgent;
+  }
+
+  &.hig__banner--warning {
+    border-color: $banner-border-color--warning;
+    background-color: $banner-background-color--warning;
+  }
+
+  &.hig__banner--primary {
+    border-color: $banner-border-color--primary;
+    background-color: $banner-background-color--primary;
+  }
+
+  &.hig__banner--complete {
+    border-color: $banner-border-color--complete;
+    background-color: $banner-background-color--complete;
+  }
+}
+
+.hig__banner__content {
+  display: flex;
+  flex-grow: 1;
+  flex-shrink: 1;
+  align-items: center;
+  justify-content: space-between;
+  margin: $banner-message-padding-y 0 0 $banner-content-padding-x;
+  // Required for proper calculation of the width
+  overflow: hidden;
+
+  .hig__banner--wrap-content & {
+    flex-direction: column;
+    align-items: flex-start;
+  }
+}
+
+.hig__banner__notification {
+  flex-basis: 50%;
+  min-width: $banner-notification-width-min;
+  max-width: $banner-notification-width-max;
+  padding-bottom: $banner-message-padding-y;
+  margin: 0 auto 0 0;
+
+  .hig__banner--interactive & {
+    margin-right: $banner-content-spacing-min;
+  }
+
+  .hig__banner--wrap-content & {
+    margin-right: 0;
+  }
+}
+
+.hig__banner__dismiss-button {
+  display: flex;
+  align-items: center;
+  flex-grow: 0;
+  flex-shrink: 0;
+  padding-right: $banner-dismiss-padding-right;
+
+  .hig__banner--interactive & {
+    padding-right: $banner-dismiss-padding-right--interactive;
+  }
+
+  .hig__banner--wrap-content & {
+    align-items: flex-start;
+    padding-top: $banner-dismiss-padding-top--wrapping;
+  }
+}
+
+.hig__banner__icon-background {
+  display: flex;
+  flex-grow: 0;
+  flex-shrink: 0;
+  align-items: center;
+  justify-content: center;
+  width: $banner-wrapper-min-height;
+  margin: 0;
+
+  svg * {
+    fill: color(hig-white);
+  }
+
+  .hig__banner--urgent & {
+    background-color: $banner-icon-background-color--urgent;
+  }
+
+  .hig__banner--warning & {
+    background-color: $banner-icon-background-color--warning;
+  }
+
+  .hig__banner--primary & {
+    background-color: $banner-icon-background-color--primary;
+  }
+
+  .hig__banner--complete & {
+    background-color: $banner-icon-background-color--complete;
+  }
+}
+
+.hig__banner__interactions-wrapper {
+  flex-shrink: 0;
+  padding-right: $banner-action-spacing-x;
+  padding-bottom: $banner-interactions-wrapper-padding-y;
+}
