+++ packages/react/src/elements/components/Banner/placements.test.js
@@ -1,0 +1,11 @@
+import * as placements from "./placements";
+
+describe("banner/placements", () => {
+  it("has an array of available placements", () => {
+    expect(placements).toHavePropertyOfConstants("AVAILABLE_PLACEMENTS");
+  });
+
+  it("has constants for placements", () => {
+    expect(placements).toHavePropertyOfConstants("placements");
+  });
+});
