+++ packages/react/src/elements/components/Banner/tokens.scss
@@ -1,0 +1,45 @@
+@import "~@hig/styles/mixins/colors.scss";
+@import "~@hig/styles/mixins/spacing.scss";
+
+/* Dimensions */
+
+$banner-action-padding-y: spacing(xxs);
+$banner-action-spacing-x: spacing(l);
+$banner-content-padding-x: spacing(m);
+// Minimum space between the notification and actions
+$banner-content-spacing-min: 125px;
+$banner-dismiss-icon-button-width: 36px;
+$banner-dismiss-icon-width: 10px;
+$banner-dismiss-spacing-right: 50px;
+// Reference from typography tokens
+$banner-message-font-size: 14px;
+$banner-message-line-height: 20px;
+$banner-notification-width-max: 900px;
+$banner-notification-width-min: 200px;
+$banner-wrapper-min-height: 50px;
+
+$banner-content-spacing-y: ($banner-wrapper-min-height - $banner-message-font-size) / 2;
+$banner-dismiss-icon-padding: ($banner-dismiss-icon-button-width - $banner-dismiss-icon-width) / 2;
+$banner-dismiss-padding-right: $banner-dismiss-spacing-right - $banner-dismiss-icon-padding;
+$banner-dismiss-spacing-right--interactive: $banner-action-spacing-x;
+$banner-dismiss-padding-right--interactive: $banner-dismiss-spacing-right--interactive - $banner-dismiss-icon-padding;
+$banner-dismiss-padding-top--wrapping: $banner-content-spacing-y - $banner-dismiss-icon-padding;
+$banner-interactions-wrapper-padding-y: $banner-content-spacing-y - $banner-action-padding-y;
+$banner-message-padding-y: ($banner-wrapper-min-height - $banner-message-line-height) / 2;
+
+/* Colors */
+
+$banner-border-color--urgent: color(hig-salmon-40);
+$banner-border-color--warning: color(hig-gold-40);
+$banner-border-color--primary: color(hig-blue-30);
+$banner-border-color--complete: color(hig-green-40);
+
+$banner-background-color--urgent: color(hig-salmon-20);
+$banner-background-color--warning: color(hig-gold-20);
+$banner-background-color--primary: color(hig-blue-10);
+$banner-background-color--complete: color(hig-green-20);
+
+$banner-icon-background-color--urgent: color(hig-red-alert);
+$banner-icon-background-color--warning: color(hig-gold-50);
+$banner-icon-background-color--primary: color(hig-blue-50);
+$banner-icon-background-color--complete: color(hig-green-50);
