--- packages/react/src/elements/components/Banner/variables.scss
@@ -1,50 +1,0 @@
-@import "~@hig/styles/mixins/colors.scss";
-@import "~@hig/styles/mixins/spacing.scss";
-
-/*
- * Banner specific variables
- * These variables shouldn't be imported globally.
- */
-
-/* Dimensions */
-
-$action-padding-y: spacing(xxs);
-$action-spacing-x: spacing(l);
-$content-padding-x: spacing(m);
-// Minimum space between the notification and actions
-$content-spacing-min: 125px;
-$dismiss-icon-button-width: 36px;
-$dismiss-icon-width: 10px;
-$dismiss-spacing-right: 50px;
-// Reference from typography tokens
-$message-font-size: 14px;
-$message-line-height: 20px;
-$notification-width-max: 900px;
-$notification-width-min: 200px;
-$wrapper-min-height: 50px;
-
-$content-spacing-y: ($wrapper-min-height - $message-font-size) / 2;
-$dismiss-icon-padding: ($dismiss-icon-button-width - $dismiss-icon-width) / 2;
-$dismiss-padding-right: $dismiss-spacing-right - $dismiss-icon-padding;
-$dismiss-spacing-right--interactive: $action-spacing-x;
-$dismiss-padding-right--interactive: $dismiss-spacing-right--interactive - $dismiss-icon-padding;
-$dismiss-padding-top--wrapping: $content-spacing-y - $dismiss-icon-padding;
-$interactions-wrapper-padding-y: $content-spacing-y - $action-padding-y;
-$message-padding-y: ($wrapper-min-height - $message-line-height) / 2;
-
-/* Colors */
-
-$border-color--error: color(hig-salmon-40);
-$border-color--warning: color(hig-gold-40);
-$border-color--primary: color(hig-blue-30);
-$border-color--success: color(hig-green-40);
-
-$background-color--error: color(hig-salmon-20);
-$background-color--warning: color(hig-gold-20);
-$background-color--primary: color(hig-blue-10);
-$background-color--success: color(hig-green-20);
-
-$icon-background-color--error: color(hig-red-alert);
-$icon-background-color--warning: color(hig-gold-50);
-$icon-background-color--primary: color(hig-blue-50);
-$icon-background-color--success: color(hig-green-50);
