+++ packages/react/support/setupTests.js
@@ -1,5 +1,5 @@
 import "jest-enzyme"; // eslint-disable-line import/no-extraneous-dependencies
 import "jest-extended";
-import "./test/matchers/toBeAnObjectOfConstants";
+import "./test/matchers/toBeConstants";
 import "./test/matchers/toImplementHIGInterfaceOf";
 import "./test/matchers/toHavePropTypesAndDocgenInfo";
